requirements.txt
flask
pandas
requests
python-Levenshtein
openpyxl
gunicorn